backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
google-generativeai==0.8.3
qdrant-client==1.11.0
sentence-transformers==3.0.1
supabase==2.7.0
python-dotenv==1.0.1
langgraph==0.1.17
langchain-core==0.2.25
typing-extensions>=4.7.0